Reorganisation of the Android sources
