+++ android/src/main/java/com/dooboolab/fluttersound/TrackPlayer.java
@@ -19,68 +19,27 @@ package com.dooboolab.fluttersound;
 
 
 
-import android.Manifest;
-import android.app.Activity;
-import android.content.BroadcastReceiver;
-import android.content.ComponentName;
 import android.content.Context;
-import android.content.Intent;
-import android.content.pm.PackageManager;
 import android.media.AudioDeviceInfo;
-import android.media.MediaPlayer;
-import android.media.MediaRecorder;
 import android.media.AudioManager;
 import android.os.Build;
-import android.os.Bundle;
 import android.os.Handler;
-import android.os.RemoteException;
-import android.os.SystemClock;
-import android.support.v4.media.MediaBrowserCompat;
 import android.support.v4.media.MediaMetadataCompat;
-import android.support.v4.media.session.MediaControllerCompat;
-import android.support.v4.media.session.MediaSessionCompat;
 import android.support.v4.media.session.PlaybackStateCompat;
 import android.util.Log;
 
-import androidx.annotation.NonNull;
 import androidx.arch.core.util.Function;
-import androidx.core.app.ActivityCompat;
-
-import android.media.AudioFocusRequest;
-
-import java.io.*;
-
-import org.json.JSONException;
-import org.json.JSONObject;
 
 import java.io.File;
 import java.io.FileOutputStream;
-import java.util.ArrayList;
 import java.util.HashMap;
 import java.util.Map;
 import java.util.Timer;
 import java.util.TimerTask;
 import java.util.concurrent.Callable;
-import java.util.concurrent.ExecutorService;
-import java.util.concurrent.Executors;
 
-import io.flutter.embedding.engine.FlutterEngine;
-import io.flutter.embedding.engine.plugins.FlutterPlugin;
-import io.flutter.embedding.engine.plugins.activity.ActivityAware;
-import io.flutter.embedding.engine.plugins.activity.ActivityPluginBinding;
-import io.flutter.plugin.common.BinaryMessenger;
 import io.flutter.plugin.common.MethodCall;
-import io.flutter.plugin.common.MethodChannel;
-import io.flutter.plugin.common.MethodChannel.MethodCallHandler;
 import io.flutter.plugin.common.MethodChannel.Result;
-
-import java.util.concurrent.Callable;
-
-import io.flutter.plugin.common.PluginRegistry;
-import io.flutter.plugin.common.PluginRegistry.Registrar;
-import com.dooboolab.fluttersound.FlutterSoundPlayer;
-import com.dooboolab.fluttersound.MediaBrowserHelper;
-import com.dooboolab.fluttersound.Track;
 
 
 //-----------------------------------------------------------------------------------------------------------------------------
@@ -108,7 +67,7 @@ public class TrackPlayer extends FlutterSoundPlayer
 	void initializeFlautoPlayer( final MethodCall call, final Result result )
 	{
 		//super.initializeFlautoPlayer( call, result );
-		audioManager = ( AudioManager ) FlautoPlayerPlugin.androidContext.getSystemService ( Context.AUDIO_SERVICE );
+		audioManager = ( AudioManager ) FlautoPlayerManager.androidContext.getSystemService ( Context.AUDIO_SERVICE );
 		assert(Flauto.androidActivity != null);
 
 		// Initialize the media browser if it hasn't already been initialized
